+++ src/main/java/kr/or/connect/reservation/service/DisplayInfoService.java
@@ -5,6 +5,6 @@ import java.util.List;
 import kr.or.connect.reservation.dto.DisplayInfo;
 
 public interface DisplayInfoService {
-	public List<DisplayInfo> getDisplayInfoById(int categoryId, int start);
-	public List<DisplayInfo> getDisplayInfo(int start);
+	public List<DisplayInfo> getDisplayInfoById(Long categoryId, Long start);
+	public List<DisplayInfo> getDisplayInfo(Long start);
 }
